requirements.txt
str2bool
python-dateutil
jsonschema
PyYAML
prometheus_client
requests